// reference/Texas-FrontEnd-TexasInformativeWebsite-release/src/types/resources.ts
export interface HeaderResourcesProps {
  ourstory: string;
  Menu: string;
  Coupons: string;
  Locations: string;
  Contactus: string;
  OrderNow: string;
  Vouchers: string;
}

export interface FooterResourcesProps {
  abouttexas: string;
  HALAL: string;
  ourstory: string;
  Menu: string;
  discovertexas: string;
  Coupons: string;
  rewards: string;
  Catring: string;
  Blog: string;
  letstalk: string;
  Contactus: string;
  FAQs: string;
  party: string;
  birthday: string;
  Footer_Workwithus: string;
  franchiseopportunities: string;
  careers: string;
  followus: string;
  Terms_Conditions: string;
  PrivacyPolicy: string;
  Footer_rights: string;
  appData: string;
  Locations: string;
  OrderNow: string;
  Nutritional_Information: string;
  legalNotice: string;
}